+++ dst/app/routes/actions.js
@@ -51,7 +51,7 @@ actionsRouter.get('', permitScopes_1.default(['actions.*', 'actions.read']), rat
 /**
  * チケット印刷アクション追加
  */
-actionsRouter.post('/print/ticket', permitScopes_1.default(['actions', 'actions.printTicket.*']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+actionsRouter.post('/print/ticket', permitScopes_1.default(['actions.*', 'actions.printTicket.*']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const ticket = {
             ticketToken: req.body.ticketToken
@@ -67,7 +67,7 @@ actionsRouter.post('/print/ticket', permitScopes_1.default(['actions', 'actions.
 /**
  * チケット印刷アクション検索
  */
-actionsRouter.get('/print/ticket', permitScopes_1.default(['actions.printTicket.*']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+actionsRouter.get('/print/ticket', permitScopes_1.default(['actions.*', 'actions.printTicket.*']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const actions = yield new cinerino.repository.Action(mongoose.connection).searchPrintTicket({
             agentId: req.user.sub,